Review Page Form
I made a form for the reveiw page. I also made a css so that it would work right. You can type your name, dorp down menu for rating slection up to 5 stars and the leave a review. There is also a submit button. Couldnt fully test it since Idk what we are even reviewing.