Syling and content Review page
So I completely redid the review page conten since the review were already written and not ment to submit. I added posters to every review which the client did not provide but asked for in the sigth plan. I addded h1,h2,h3, and p styling colors and a color for the body background so I could see how everything pops. If needed to chang we can. if you want to center text on other pages it wil be best to add the h1, h2, h3, and p styling i made to the div for my page. This way the flow set up still runs but that is totaly up to the lead.